CEMT Sync Code enhancement, CEMT databases will generate a unique sync key that will be propagated to OD databases when syncs occur. Reviewed by: Allen Each CEMT database can contain a randomly generated 10-character alpha-numeric code which can contain any upper or lower case letter, or any number from 0 to 9.  It is obtained when opening the CEMT program the first time.  It is unique per CEMT database. Example: TrE9P4n7XC
This code is displayed in two places.  First, in the CEMT in the security setup window.  Second, in OD proper in the Miscellaneous setup window.
The code can be cleared from OD proper, but cannot be cleared by the CEMT.  Once it is generated on login to the CEMT the first time it is permanent.
When Syncing anything from the CEMT to any other OD database it will check the sync codes to see if they match. 
If they match, the CEMT will sync data appropriately.
If they don't match, two things could happen: One, if the OD database has a sync code already the CEMT information will not be synced to that OD database.  Second, if the OD database has no sync code it will receive the CEMT's sync code and syncing will commence.  What this means is that if an OD database has a sync code it can only be synced to by a specific CEMT database.
In the Miscellaneous setup window in OD proper the code is displayed and there is also a Clear button.  The Clear button will remove the code from the OD database and allow syncing from any CEMT once more.

git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@14223 098d0599-7f2b-0410-95a5-ad30ac85393a